--- src/components/admin/main/information/MainHeroDataGrid.tsx
@@ -1,5 +1,0 @@
-const MainInformationDataGrid = () => {
-  return <div>창업 정보</div>;
-};
-
-export default MainInformationDataGrid;
