+++ src/components/admin/main/information/MainInformationDataGrid.tsx
@@ -1,0 +1,7 @@
+import Box from '@mui/material/Box';
+
+const MainInformationDataGrid = () => {
+  return <Box sx={{ padding: 2 }}>창업정보</Box>;
+};
+
+export default MainInformationDataGrid;
